backend/functions/requirements.txt
PyJwt==1.7.1
pynamodb==4.3.2
requests==2.23.0